Add shared template name validator
