+++ app/routes.py
@@ -58,6 +58,7 @@ from app.utils import (
 from app.utils.db import SessionLocal
 #from app.models.log import Log
 from app.utils.scheduling import log_cache, log_cache_lock
+from app.utils.validators import validate_template_name
 
 def restart_server():
     print("Restarting server...")
@@ -70,40 +71,6 @@ def restart_server():
     restart_thread = Thread(target=delayed_restart)
     restart_thread.start()
 
-# todo: add this to utils so it is not duplicated in utils/video_archiver.py
-def validate_template_name(template_name: str):
-    if template_name is None or not isinstance(template_name, str):
-        return None
-
-    # Strict whitelist of allowed characters
-    allowed_chars = set('abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_-.')
-
-    # Check if all characters are in the allowed set
-    if not all(char in allowed_chars for char in template_name):
-        return None
-
-    # Check length
-    if len(template_name) == 0 or len(template_name) > 32:
-        return None
-
-    # Ensure the name doesn't start or end with a dash or underscore
-    if template_name[0] in '-_.' or template_name[-1] in '-_.':
-        return None
-    if '..' in template_name:
-        return None
-    if '--' in template_name:
-        return None
-    if '__' in template_name:
-        return None
-
-    # Use secure_filename as an additional safety measure
-    sanitized_name = secure_filename(template_name)
-
-    # Ensure secure_filename didn't change the name (which would indicate it found something suspicious)
-    if sanitized_name != template_name:
-        return None
-
-    return sanitized_name
 
 
 class TemplateName:
