+++ app/utils/video_archiver.py
@@ -8,7 +8,7 @@ import subprocess
 import tempfile
 import time
 
-from werkzeug.utils import secure_filename
+from .validators import validate_template_name
 
 from app.config import (
     MAX_COMPRESSED_VIDEO_AGE,
@@ -22,42 +22,6 @@ from app.config import (
 )
 
 from .template_manager import get_templates
-
-
-# TODO: move this to utils so it is not duplicated in routes.py
-def validate_template_name(template_name: str):
-    if template_name is None or not isinstance(template_name, str):
-        return None
-
-    # Strict whitelist of allowed characters
-    allowed_chars = set('abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_-.')
-
-    # Check if all characters are in the allowed set
-    if not all(char in allowed_chars for char in template_name):
-        return None
-
-    # Check length
-    if len(template_name) == 0 or len(template_name) > 32:
-        return None
-
-    # Ensure the name doesn't start or end with a dash or underscore
-    if template_name[0] in '-_.' or template_name[-1] in '-_.':
-        return None
-    if '..' in template_name:
-        return None
-    if '--' in template_name:
-        return None
-    if '__' in template_name:
-        return None
-
-    # Use secure_filename as an additional safety measure
-    sanitized_name = secure_filename(template_name)
-
-    # Ensure secure_filename didn't change the name (which would indicate it found something suspicious)
-    if sanitized_name != template_name:
-        return None
-
-    return sanitized_name
